# Per default refetch the chain tip at most every 20 seconds