--- a/pycardano/backend/ogmios.py
+++ b/pycardano/backend/ogmios.py
@@ -1,8 +1,8 @@
 import json
+import time
 from datetime import datetime, timezone
 from enum import Enum
 from typing import Any, Dict, List, Optional, Tuple, Union
-import time
 
 import cbor2
 import requests