@nielstron 2023-04-16
Using the opshin-starter-kit as an example

Before
```
(opshin-starter-kit-py3.8) ➜  opshin-starter-kit git:(main) ✗ time python3 src/off_chain/make_vest.py owner beneficiary
transaction id: 4a26f76b7354d0d0f1c232dc187822701a1fe3faa565ffdc94a08ff9282eb51f
Cardanoscan: https://preview.cardanoscan.io/transaction/4a26f76b7354d0d0f1c232dc187822701a1fe3faa565ffdc94a08ff9282eb51f
python3 src/off_chain/make_vest.py owner beneficiary  0.60s user 0.08s system 1% cpu 43.842 total
```

After
```
python3 src/off_chain/make_vest.py owner beneficiary  0.28s user 0.05s system 5% cpu 6.408 total
(opshin-starter-kit-py3.8) ➜  opshin-starter-kit git:(main) ✗ time python3 src/off_chain/make_vest.py owner beneficiary
transaction id: 7042a915046c87c8cd1d7b953f09e3738fb3c93c585730f9546b7000fc32d018
Cardanoscan: https://preview.cardanoscan.io/transaction/7042a915046c87c8cd1d7b953f09e3738fb3c93c585730f9546b7000fc32d018
python3 src/off_chain/make_vest.py owner beneficiary  0.27s user 0.04s system 3% cpu 7.921 total
```

This is a speed up from 40s to 7s and the script only builds a simple transaction!