@nielstron 2023-04-16
An optimal design would have a few threads that asynchronousmy fetch the chain tip, the protocol params, the epoch and such and fetch it. Then access to those values is instant and they are almost never outdated.