@cffls 2023-04-17
The test failed. I had a fix but can't push to your fork. Could you pull [the fix](https://github.com/Python-Cardano/pycardano/tree/feat/cache_chain_tip) into your fork?